Gadgetron 2p5; together with GtCloud paper
Former-commit-id: 83ca85aef1d1f23290bb2b20eac9030f509f07cd

+++ apps/standalone/cpu/registration/3d/register_CK_3d.cpp
@@ -57,8 +57,8 @@ int main(int argc, char** argv)
     return 1;
   }
   
-  unsigned int num_fixed_dims = fixed_image->get_number_of_dimensions();
-  unsigned int num_moving_dims = moving_image->get_number_of_dimensions();
+  unsigned long long num_fixed_dims = fixed_image->get_number_of_dimensions();
+  unsigned long long num_moving_dims = moving_image->get_number_of_dimensions();
 
   if( !(num_fixed_dims == 3 || num_fixed_dims == 4)  ){
     cout << endl << "The fixed image is not three- or four-dimensional. Quitting!\n" << endl;
