+++ apps/standalone/gpu/mri/nfft/2d/main_nffth.cpp
@@ -97,7 +97,7 @@ int main( int argc, char** argv)
   delete timer;
   
   // Setup resulting image array
-  vector<unsigned int> image_dims = to_std_vector(matrix_size);
+  vector<unsigned long long> image_dims = to_std_vector(matrix_size);
   cuNDArray<_complext> image; image.create(&image_dims);
   
   // Initialize plan
