+++ apps/standalone/gpu/mri/sense/noncartesian/radial/2d_golden_ratio/main_gpbb.cpp
@@ -29,7 +29,7 @@ typedef reald<_real,2>::Type _reald2;
 boost::shared_ptr< cuNDArray<_complext> >
 upload_data( unsigned int reconstruction, unsigned int samples_per_reconstruction, unsigned int total_samples_per_coil, unsigned int num_coils, hoNDArray<_complext> *host_data )
 {
-  vector<unsigned int> dims; dims.push_back(samples_per_reconstruction); dims.push_back(num_coils);
+  vector<unsigned long long> dims; dims.push_back(samples_per_reconstruction); dims.push_back(num_coils);
   cuNDArray<_complext> *data = new cuNDArray<_complext>(); data->create( &dims );
   for( unsigned int i=0; i<num_coils; i++ )
     cudaMemcpy( data->get_data_ptr()+i*samples_per_reconstruction,
@@ -163,7 +163,7 @@ int main(int argc, char** argv)
   boost::shared_ptr< cuNDArray<_complext> > csm = estimate_b1_map<_real,2>( acc_images.get() );
   E->set_csm(csm);
 
-  std::vector<unsigned int> reg_dims = to_std_vector(matrix_size);
+  std::vector<unsigned long long> reg_dims = to_std_vector(matrix_size);
   cuNDArray<_complext> _reg_image = cuNDArray<_complext>(&reg_dims);
   E->mult_csm_conj_sum( acc_images.get(), &_reg_image );
 
@@ -186,7 +186,7 @@ int main(int argc, char** argv)
   precon_weights.reset();
   csm.reset();
 
-  boost::shared_ptr< std::vector<unsigned int> > recon_dims( new std::vector<unsigned int> );
+  boost::shared_ptr< std::vector<unsigned long long> > recon_dims( new std::vector<unsigned long long> );
   *recon_dims = to_std_vector(matrix_size); recon_dims->push_back(frames_per_reconstruction);
 
   delete timer;
@@ -195,7 +195,7 @@ int main(int argc, char** argv)
   // Setup radial SENSE reconstructions
   //
 
-  vector<unsigned int> data_dims;
+  vector<unsigned long long> data_dims;
   data_dims.push_back(samples_per_reconstruction); data_dims.push_back(num_coils);
 
   E->set_domain_dimensions(recon_dims.get());
@@ -229,7 +229,7 @@ int main(int argc, char** argv)
   unsigned int num_reconstructions = num_profiles / profiles_per_reconstruction;
 
   // Allocate space for result
-  std::vector<unsigned int> res_dims = to_std_vector(matrix_size);
+  std::vector<unsigned long long> res_dims = to_std_vector(matrix_size);
   res_dims.push_back(frames_per_reconstruction*num_reconstructions);
   cuNDArray<_complext> result = cuNDArray<_complext>(&res_dims);
 
@@ -258,7 +258,7 @@ int main(int argc, char** argv)
       solve_result = solver.solve(data.get());
     }
 
-    vector<unsigned int> tmp_dims = to_std_vector(matrix_size); tmp_dims.push_back(frames_per_reconstruction);
+    vector<unsigned long long> tmp_dims = to_std_vector(matrix_size); tmp_dims.push_back(frames_per_reconstruction);
     cuNDArray<_complext> tmp(&tmp_dims, result.get_data_ptr()+reconstruction*prod(matrix_size)*frames_per_reconstruction );
 
     // Copy sbresult to result (pointed to by tmp)
