+++ gadgets/mri_core/MRIImageWriter.cpp
@@ -53,7 +53,7 @@ int MRIImageWriter<T>::write(ACE_SOCK_Stream* sock, ACE_Message_Block* mb)
 		GADGET_DEBUG2("Number of header elements %d is inconsistent with number of elements in NDArray %d\n",expected_elements, datamb->getObjectPtr()->get_number_of_elements());
 		GADGET_DEBUG2("Header dimensions: %d, %d, %d\n",imagemb->getObjectPtr()->matrix_size[0],imagemb->getObjectPtr()->matrix_size[1],imagemb->getObjectPtr()->matrix_size[2]);
 		GADGET_DEBUG2("Number of array dimensions: %d:\n", datamb->getObjectPtr()->get_number_of_dimensions());
-		for (unsigned int i = 0; i < datamb->getObjectPtr()->get_number_of_dimensions(); i++) {
+		for (unsigned long long i = 0; i < datamb->getObjectPtr()->get_number_of_dimensions(); i++) {
 			GADGET_DEBUG2("Dimensions %d: %d\n", i, datamb->getObjectPtr()->get_size(i));
 		}
 		return -1;
