+++ gadgets/mri_core/NoiseAdjustGadget_unoptimized.cpp
@@ -127,7 +127,7 @@ int NoiseAdjustGadget_unoptimized
 		noise_dwell_time_us_ = m1->getObjectPtr()->sample_time_us;
 		//If noise covariance matrix is not allocated
 		if (noise_covariance_matrix_.get_number_of_elements() != channels*channels) {
-			std::vector<unsigned int> dims(2, channels);
+			std::vector<unsigned long long> dims(2, channels);
 			try{ noise_covariance_matrix_.create(&dims);}
 			catch (std::runtime_error &err){
 				GADGET_DEBUG_EXCEPTION(err,"Unable to allocate storage for noise covariance matrix\n");
