+++ gadgets/sense/gpuCgKtSenseGadget.cpp
@@ -184,7 +184,7 @@ namespace Gadgetron{
     GADGET_DEBUG2("Matrix size    : [%d,%d] \n", matrix_size_[0], matrix_size_[1]);    
     GADGET_DEBUG2("Matrix size OS : [%d,%d] \n", matrix_size_os_[0], matrix_size_os_[1]);
 
-    std::vector<unsigned int> image_dims = to_std_vector(matrix_size_);
+    std::vector<unsigned long long> image_dims = to_std_vector(matrix_size_);
     image_dims.push_back(frames);
     
     E_->set_domain_dimensions(&image_dims);
@@ -265,7 +265,7 @@ namespace Gadgetron{
       *m->getObjectPtr() = j->image_headers_[frame];
       m->cont(cm);
       
-      std::vector<unsigned int> img_dims(2);
+      std::vector<unsigned long long> img_dims(2);
       img_dims[0] = matrix_size_seq_[0];
       img_dims[1] = matrix_size_seq_[1];
 
@@ -316,7 +316,7 @@ namespace Gadgetron{
     unsigned int num_rotations = num_samples / job->tra_host_->get_number_of_elements();
     unsigned int frames_per_reconstruction = job->tra_host_->get_size(1)*num_rotations;
 
-    std::vector<unsigned int> dims = to_std_vector(matrix_size_os_);
+    std::vector<unsigned long long> dims = to_std_vector(matrix_size_os_);
     dims.push_back(frames_per_reconstruction); 
     dims.push_back(num_coils);
