+++ gadgets/sense/gpuCgSenseGadget.cpp
@@ -189,7 +189,7 @@ namespace Gadgetron{
       matrix_size_reported_ = true;
     }
 
-    std::vector<unsigned int> image_dims = to_std_vector(matrix_size_);
+    std::vector<unsigned long long> image_dims = to_std_vector(matrix_size_);
     image_dims.push_back(frames);
     
     E_->set_domain_dimensions(&image_dims);
@@ -273,7 +273,7 @@ namespace Gadgetron{
       *m->getObjectPtr() = j->image_headers_[frame];
       m->cont(cm);
       
-      std::vector<unsigned int> img_dims(2);
+      std::vector<unsigned long long> img_dims(2);
       img_dims[0] = matrix_size_seq_[0];
       img_dims[1] = matrix_size_seq_[1];
