+++ gadgets/sense/gpuGenericSensePrepGadget.h
@@ -113,9 +113,9 @@ namespace Gadgetron{
     boost::shared_array< cuSenseBuffer<float,2> > acc_buffer_;
     boost::shared_array< cuSenseBufferCg<float,2> > acc_buffer_cg_;
 
-    std::vector<unsigned int> fov_;
-    std::vector<unsigned int> image_dimensions_;
-    std::vector<unsigned int> image_dimensions_recon_;
+    std::vector<unsigned long long> fov_;
+    std::vector<unsigned long long> image_dimensions_;
+    std::vector<unsigned long long> image_dimensions_recon_;
     uintd2 image_dimensions_recon_os_;
 
     boost::shared_array< ACE_Message_Queue<ACE_MT_SYNCH> > frame_readout_queue_;
