+++ gadgets/sense/gpuSbSenseGadget.cpp
@@ -227,7 +227,7 @@ namespace Gadgetron{
       GADGET_DEBUG2("Matrix size    : [%d,%d] \n", matrix_size_[0], matrix_size_[1]);
       GADGET_DEBUG2("Matrix size OS : [%d,%d] \n", matrix_size_os_[0], matrix_size_os_[1]);
 
-      std::vector<unsigned int> image_dims = to_std_vector(matrix_size_);
+      std::vector<unsigned long long> image_dims = to_std_vector(matrix_size_);
       image_dims.push_back(frames);
       
       E_->set_domain_dimensions(&image_dims);
@@ -382,7 +382,7 @@ namespace Gadgetron{
       m->getObjectPtr()->matrix_size[1] = matrix_size_seq_[1];      
       m->cont(cm);
       
-      std::vector<unsigned int> img_dims(2);
+      std::vector<unsigned long long> img_dims(2);
       img_dims[0] = matrix_size_seq_[0];
       img_dims[1] = matrix_size_seq_[1];
